+++ src/components/Footer.tsx
@@ -1,6 +1,6 @@
-import React, { memo, useCallback, useEffect, useRef } from "react";
+import React, { memo } from "react";
 import { ColumnProps, FooterProps } from "../..";
-import { cx, findTableByUuid } from "../util";
+import { cx } from "../util";
 
 type InnerFooterCellProps<T> = {
   width: number;
@@ -32,6 +32,7 @@ const FooterCell = memo(BaseFooterCell) as <T>(props: InnerFooterCellProps<T>) =
 type InnerFooterProps<T> = {
   uuid: string;
   rows: T[];
+  isScrollHorizontal: boolean;
   columns: ColumnProps<T>[];
   pixelWidths: number[];
   className?: string;
@@ -47,12 +48,10 @@ function Footer<T>({
   pixelWidths,
   sticky,
   className,
+  isScrollHorizontal,
   style: footerStyle = {},
   component: Component
 }: InnerFooterProps<T>) {
-  const scroll = useRef(false);
-  const ref = useRef<HTMLDivElement>(null);
-
   // constants
   const hasFooter = !!Component || columns.some(c => !!c.footer);
   const style: React.CSSProperties = {
@@ -64,60 +63,14 @@ function Footer<T>({
     style.minWidth = 0;
   }
 
-  // functions
-  const onScroll = useCallback(
-    (target: HTMLElement | null, element: HTMLElement | null) => {
-      if (scroll.current || !element || !sticky) {
-        return;
-      }
-
-      const left = target?.scrollLeft || 0;
-      scroll.current = true;
-      element.scrollLeft = left;
-      setTimeout(() => {
-        scroll.current = false;
-      }, 0);
-    },
-    [uuid, sticky]
-  );
-
-  const listener = useCallback(
-    (ev: Event) => {
-      onScroll(ev.target as HTMLDivElement, ref.current);
-    },
-    [onScroll]
-  );
-
-  // effects
-  // add scroll listener
-  useEffect(() => {
-    if (uuid) {
-      const e = findTableByUuid(uuid);
-      if (e) {
-        e.addEventListener("scroll", listener);
-      }
-    }
-
-    return () => {
-      if (uuid) {
-        const e = findTableByUuid(uuid);
-        if (e) {
-          e.removeEventListener("scroll", listener);
-        }
-      }
-    };
-  }, [uuid, listener]);
-
   // render
   if (!Component) {
     const hasFooter = columns.some(c => !!c.footer);
     return (
       <div
-        ref={ref}
         style={{ border: !hasFooter ? "none" : undefined, ...style }}
         data-footer-key={`${uuid}-footer`}
-        className={cx("rft-footer", sticky && "sticky", className)}
-        onScroll={e => onScroll(e.target as HTMLDivElement, findTableByUuid(uuid))}
+        className={cx("rft-footer", sticky && "sticky", isScrollHorizontal && "scroll", className)}
       >
         <div className="rft-row">
           {columns.map((c, i) => (
@@ -136,11 +89,9 @@ function Footer<T>({
 
   return (
     <div
-      ref={ref}
       style={style}
       data-footer-key={`${uuid}-footer`}
-      className={cx("rft-footer", sticky && "sticky", className)}
-      onScroll={e => onScroll(e.target as HTMLDivElement, findTableByUuid(uuid))}
+      className={cx("rft-footer", sticky && "sticky", isScrollHorizontal && "scroll", className)}
     >
       <Component rows={rows} widths={pixelWidths} />
     </div>